Moved package 'rax' to nest under 'backends' as it should be.
